+++ include/unity/scopes/Object.h
@@ -57,7 +57,7 @@ public:
     \brief Returns the category of the target object of this proxy.
     \return The category of the target of the proxy.
     */
-    virtual std::string category() = 0;
+    virtual std::string target_category() = 0;
 
     /**
     \brief Returns the timeout in milliseconds if this proxy is a twoway proxy.