+++ src/scopes/internal/RegistryImpl.cpp
@@ -71,6 +71,11 @@ MetadataMap RegistryImpl::list_if(std::function<bool(ScopeMetadata const& item)>
     return matching_entries;
 }
 
+void RegistryImpl::set_list_update_callback(std::function<void()> callback)
+{
+    fwd()->set_list_update_callback(callback);
+}
+
 MWRegistryProxy RegistryImpl::fwd()
 {
     return dynamic_pointer_cast<MWRegistry>(proxy());