+++ test/gtest/scopes/internal/RegistryObject/RegistryObject_test.cpp
@@ -53,7 +53,7 @@ public:
 
     MOCK_METHOD0(identity, std::string());
 
-    MOCK_METHOD0(category, std::string());
+    MOCK_METHOD0(target_category, std::string());
 
     MOCK_METHOD0(timeout, int64_t());
 
@@ -164,7 +164,7 @@ protected:
         exec_data.confinement_profile = confinement_profile;
         exec_data.timeout_ms = 1500;
 
-        registry.reset(new RegistryObject(*death_observer(), executor));
+        registry.reset(new RegistryObject(*death_observer(), executor, nullptr));
         registry->add_local_scope("scope-id", meta, exec_data);
         registry->locate("scope-id");
         EXPECT_TRUE(registry->is_scope_running("scope-id"));