+++ RELEASE_NOTES.md
@@ -10,11 +10,14 @@ Changes in version 0.4.8
   - Changes to departments API: SearchReply::register_departments() method now takes parent and current department argument,
     and uses Department::SCPtr for both. SearchListenerBase::push() method for departments got changed to match as well. Removed
     constructors of Department and added static create() methods instead. Changed DepartmentList to hold Department
-    pointers instead of values.
+    pointers instead of values. Changed Department::set_has_subdepartments() method to take bool value (true by default).
+
+  - Changed parameter type for pushing categories on SearchListenerBase to Category::SCPtr const&.
 
 Changes in version 0.4.7
 ========================
   - Implemented RatingFilter and RadioButtonsFilter.
+
   - changed create() methods of OptionSelectorFilter and RangeInputFilter to return unique_ptr (UPtr)
     instead of shared pointers.